[BUGFIX] Stop trying to sync deleted files
If a file was deleted in pixx.io, but the "delete" function of
the sync command is disabled, it tries to keep syncing the same
files all the time, because it does not remember that they are
deleted.

To fix this, the table `sys_file_metadata` has a new field
`pixxio_deleted`. It is set to "1" if a file is deleted in
pixx.io but should remain in TYPO3. Once it is set, the sync
command will ignore it.